+++ climt/_components/rrtmg/lw/component.py
@@ -8,7 +8,8 @@ from ...._core import (
 from numpy import pi as PI
 from ..rrtmg_common import (
     rrtmg_cloud_overlap_method_dict, rrtmg_cloud_props_dict,
-    rrtmg_cloud_ice_props_dict, rrtmg_cloud_liquid_props_dict)
+    rrtmg_cloud_ice_props_dict, rrtmg_cloud_liquid_props_dict,
+    rrtmg_random_number_dict)
 import logging
 try:
     from . import _rrtmg_lw
@@ -166,7 +167,7 @@ class RRTMGLongwave(TendencyComponent):
             cloud_liquid_water_properties='radius_dependent_absorption',
             calculate_interface_temperature=True,
             mcica=False,
-            random_number_generator=0,
+            random_number_generator='mersenne_twister',
             **kwargs):
         """
 
@@ -227,10 +228,10 @@ class RRTMGLongwave(TendencyComponent):
                 * mcica = True: use the McICA version of the longwave component of RRTMG
                 * mcica = False: use the nomcica version of the longwave component of RRTMG
 
-            random_number_generator (int):
+            random_number_generator (string):
                 Different methods of generating random numbers for McICA.
-                * random_number_generator = 0: kissvec
-                * random_number_generator = 1: Mersenne Twister
+                * :code:`kissvec`
+                * :code:`mersenne_twister`
 
         .. _[Ebert and Curry 1992]:
             http://onlinelibrary.wiley.com/doi/10.1029/91JD02472/abstract
@@ -251,7 +252,8 @@ class RRTMGLongwave(TendencyComponent):
         self._mcica = mcica
         if mcica:
             self._permute_seed = None
-            self._random_number_generator = random_number_generator
+            self._random_number_generator = rrtmg_random_number_dict[
+                random_number_generator.lower()]
             if type(cloud_overlap_method) is str:
                 if cloud_overlap_method.lower() == 'clear_only':
                     logging.info(
@@ -368,7 +370,15 @@ class RRTMGLongwave(TendencyComponent):
             # radiation is called, with the same state / input properties,
             # a different result is obtained, because the wavelengths which
             # see cloud differ between each call.
-            self._permute_seed = np.random.randint(0, 500)
+            if self._random_number_generator == 0:
+                # KISS algorithm: The seed determines the number of times
+                # the random number generator is called iteratively to create a
+                # new random number. The value range of the seed is limited to
+                # avoid a performance decrease.
+                self._permute_seed = np.random.randint(0, 1024)
+            elif self._random_number_generator == 1:
+                # Mersenne Twister: Use random seed from the full 32bit range.
+                self._permute_seed = np.random.randint(0, 2**31 - 1)
 
             _rrtmg_lw.initialise_rrtm_radiation_mcica(
                 self._Cpd,